% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSquare.R
\name{chisq.pdf}
\alias{chisq.pdf}
\title{Plot the PDF of the Chi-Squared Distribution}
\usage{
chisq.pdf(params = NULL, limits = c(NULL, NULL), probs = c(NULL, NULL), ...)
}
\arguments{
\item{params}{Optional list of parameters for the distribution (e.g., degrees of freedom \code{df}).}

\item{limits}{Optional numeric vector specifying the lower and upper bounds for shading.}

\item{probs}{Optional numeric vector of probabilities to convert into bounds using the quantile function.}

\item{...}{Additional graphical parameters passed to the plotting function.}
}
\value{
A plot is produced. Invisibly returns \code{NULL}.
}
\description{
Generates a plot of the probability density function (PDF) for the chi-squared distribution, with optional shading between specified limits or quantiles.
}
\examples{
chisq.pdf(params = c(df = 10))
chisq.pdf(params = c(df = 10), limits = c(10, Inf))
chisq.pdf(params = c(df = 10), probs = c(.95, 1))
}
