% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnedecorsF.R
\name{f.pdf}
\alias{f.pdf}
\title{Plot the PDF of the F Distribution}
\usage{
f.pdf(params = NULL, limits = c(NULL, NULL), probs = c(NULL, NULL), ...)
}
\arguments{
\item{params}{Optional list of parameters for the distribution (e.g., \code{df1}, \code{df2}).}

\item{limits}{Optional numeric vector specifying the lower and upper bounds for shading.}

\item{probs}{Optional numeric vector of probabilities to convert into bounds using the quantile function.}

\item{...}{Additional graphical parameters passed to the plotting function.}
}
\value{
A plot is produced. Invisibly returns \code{NULL}.
}
\description{
Generates a plot of the probability density function (PDF) for the F distribution, with optional shaded areas between specified limits or quantiles.
}
\examples{
f.pdf(params = c(df1 = 2, df2 = 10))
f.pdf(params = c(df1 = 2, df2 = 10), limits = c(3, Inf))
f.pdf(params = c(df1 = 2, df2 = 10), probs = c(.95, 1))
}
