% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogNormal.R
\name{lnorm.cdf}
\alias{lnorm.cdf}
\title{Plot the CDF of the Log-Normal Distribution}
\usage{
lnorm.cdf(params = NULL, limits = c(NULL, NULL), probs = c(NULL, NULL), ...)
}
\arguments{
\item{params}{Optional list of parameters for the distribution (e.g., \code{meanlog}, \code{sdlog}).}

\item{limits}{Optional numeric vector of lower and upper bounds to evaluate and highlight on the CDF.}

\item{probs}{Optional numeric vector of probabilities to be converted to limits.}

\item{...}{Additional graphical parameters passed to the plotting function.}
}
\value{
A plot is produced. Invisibly returns \code{NULL}.
}
\description{
Generates a plot of the cumulative distribution function (CDF) for the log-normal distribution, with optional visual indicators of probability differences between specified limits or quantiles.
}
