% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudentizedRange.R
\name{q.cdf}
\alias{q.cdf}
\title{Plot the CDF of the Studentized Range Distribution}
\usage{
q.cdf(params = NULL, limits = c(NULL, NULL), probs = c(NULL, NULL), ...)
}
\arguments{
\item{params}{Optional list of parameters for the distribution (e.g., degrees of freedom and number of means).}

\item{limits}{Optional numeric vector specifying the lower and upper bounds for annotations.}

\item{probs}{Optional numeric vector of probabilities to convert into bounds using the quantile function.}

\item{...}{Additional graphical parameters passed to the plotting function.}
}
\value{
A plot is produced. Invisibly returns \code{NULL}.
}
\description{
Generates a plot of the cumulative distribution function (CDF) for the Studentized range distribution, optionally highlighting probability differences between specified limits or quantiles.
}
\examples{
q.cdf(params = c(nmeans = 2, df = 10))
q.cdf(params = c(nmeans = 2, df = 10), limits = c(2, Inf))
q.cdf(params = c(nmeans = 2, df = 10), probs = c(.95, 1))
}
