% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistributions-package.R
\docType{package}
\name{plotDistributions}
\alias{plotDistributions}
\alias{plotDistributions-package}
\title{plotDistributions: Visualizing Probability Distributions}
\description{
plotDistributions provides functions to plot and visualize continuous distributions.
}
\details{
plotDistributions is a small R package for plotting the probability density and 
cumulative distribution functions for various continuous distributions. 
It provides functionality to highlight probabilities and display critical values.
}
\author{
**Maintainer**: Craig Wendorf 
\email{cwendorf@uwsp.edu}
}
